refactor: :recycle: update `PrismaAccountRepository`
The Prisma account repository has been updated to use fewer lines of code by condensing database queries. This change will enhance the readability of the code and make it easier to maintain and modify in the future. The webResources.xml and mirrord.json files were also added to configure project specific settings.