docs: :memo: add swagger documentation
Added a new file to build Swagger documentation for our application. This will enable us to view, test, and interact with our API's endpoints via a web interface. Included operation details in app.controller.ts to expose them in the documentation. Also, added related dependencies like "@nestjs/swagger" for defining the Swagger documentation and a couple of OpenTelemetry packages for tracing and metrics in the API. Updated the lock file accordingly.