refactor: 🎨 update `ContinuationLocalStorageModule`
The `localstorage.module.ts` has been removed and replaced with `continuation-local-storage-module.ts` for better request tracking and metadata storage. In addition, the `swagger.ts` file was updated to include swagger themes for more readable API documentation, which required the installation of the `swagger-themes` package seen in the changes in `package.json` and `pnpm-lock.yaml`.