chore: ⬆️ update dependencies and optimize import statements
Updated versions of various dependencies in package.json to keep the project up-to-date with the latest packages. This ensures that we get the benefit of any bug fixes, security patches, and new features that these updated packages may provide.

Also did some housekeeping in the import statements by removing unnecessary whitespaces in env.ts and refactoring index.js imports in main.ts. This is aimed to improve readability and maintain the overall clean code base.

Updated the system information printing condition in main.ts to only print in production environment. This reduces unnecessary console logs in non-production environments.