Add AccountModule to Authentication module and refactor account service
Added AccountModule import to the AuthenticationModule to include services, controllers, and entities available in the AccountModule. This enhances the extensibility and modularity of the authentication system.

Also refactored the AccountService. Added a register function with policy enforcement ensuring that the username is unique, email is unique, and password is secure. This strengthens the account registration security.

Additionally, created the IdentityAndAccessModule for consolidating all modules under the identity and access boundary and refactored existing code to use this new module.

Finally, various import statements have been optimized for better code organization and readability and some code cleanups.

Also updated dependencies in package.json and added a new script to copy markdown files at build time providing a way to manage documentation more effectively. Modified the tsconfig.json to include a new path alias and a typescript transformer.