feat: ✨ add argon2 as a hashing algorithm
Implemented Argon2 as a hashing strategy in the infrastructure layer. The HashingStrategy abstract class and the Argon2HashingAlgorithm class has been defined in the security/hashing directory. The HashingOptions interface has also been defined to specify the cryptographical options required for hash generation and verification. The Argon2 library has been installed and added to the package.json. This change improves the security of password hashing within the application.