chore: :hammer: added gitleaks to pre-commit config and setup script
Updated .pre-commit-config.yaml and 00_setup_repository.sh file to add Gitleaks for better security. Gitleaks is implemented to prevent any potential leakage of sensitive data or secret keys by scanning the codebase for accidental exposure of such information. It is a preventive measure enhancing the overall security of the project.