docs: 📝 add layered structure guide for software components
A comprehensive guide describing the layered structure in software components development has been added to the faq section. This will help the team understand different types of components and where they should be used in a project. The README.md file was also updated with a list of features yet to be implemented, providing a roadmap for the development of the server application. Lastly, a new library 'identification' was added to provide functions for identifying entities in the system, contributing to a more organized codebase.