feat: ✨ add session storage capabilities and session-based authentication
Added support for storing user sessions in the database, with the introduction of a new SessionRepository. The AuthenticationService now utilizes this repository to save sessions upon successful authentication. The existence check method was also added to the WriteRepository to simplify the save operation.

Additionally, the refactoring of tokens and payloads into classes such as AccessToken and RefreshToken was did, providing more explicit typing and easing potential future scalability. Adaptations were made in TokenManagement methods plus api documentation to accommodate these changes.

Overall, this commit brings in session-based user authentication for better security and compliance. Logging utility was also updated to secure sensitive data.