refactor: ♻️ pull method in data-sync classes
Updated the pull method in data-synchronization and data-source classes to no longer require a parameter. This will allow for a simpler usage of these methods and will eliminate any potential errors that could be caused by an incorrect or missing argument. The changes are part of an overall effort to improve our data synchronization process and to make our codebase more intuitive to work with.