feat: ✨ add data seeders for account and product
Added data seeders for the account and product entities. This includes seeding logic and related services. The seeders use Faker library to generate realistic account and product data. This update will significantly improve the environment setup for development and testing by pre-populating the database with realistic data. The change was necessitated by a need to test database-related functionality in the absence of actual data.