feat: ✨ add profile seeder and related updates
This commit introduces a new ProfileSeeder in the profile infrastructure. This was necessary to generate fake data for testing profile related features. It also updates the bootstrap.ts file to include ProfileSeeder in the list of providers and run it as part of seeding process.

Additionally, a new file profile-metadata.ts has been added to establish the necessary fields for profile related information. A new run configuration for Database Studio has also been introduced in the .run folder. This would help in exploring and manipulating database in a visual interface during development.

This commit enhances the testing and development process by providing necessary tools and data setup.