feat: ✨ Add discovery module for dynamic metadata retrieval
This commit introduces a discovery module which allows for metadata retrieval from NestJS components. This service will facilitate dynamic calculations and decision-making based on the metadata available at runtime. Including this allows us to reduce hardcoding of data path and dependencies, making the application more flexible. All the changes are relevant to the discovery module and metadata utilization.

Changes are done in the following files:
- Added new discovery-service.ts file to define and implement discovery service.
- Added new discovery-interface.ts to define various interfaces used in discovery service.
- Added new discovery-module.ts to wrap discovery service into a module for easy import and use in other modules.
- Added and updated stripe-decorator.ts and make-injectable-decorator.ts for supporting new discovery related functionality.
- Updated openapi3.json for documenting the API changes related to this new feature.