refactor: ♻️ storage functionality and initiate refactor towards blob storage
This update contains removals, additions, and changes related to a refactoring effort for our storage functionality. The storage functionality was removed and a new object storage abstraction module named "Blobber" is introduced to replace it. In addition, persistence models like write-model and update-model have been moved to the new "Storr" library. "Update-model" was also renamed to README.md in the new "Blobber" directory. These changes are made to improve the modularity and clarity of our codebase. Furthermore, "Geoseeker" and unidentified libraries have been added. This update also sees a renaming of several files and directories to align with the new structure. Finally, references to these files and directories have been updated across the project.