ci: 💚 add `postgres`, `nats` and `redis` for deploy job
This commit includes a new Github action to build a Docker image for a TypeScript application and publish it with specific tags. It also updates an existing workflow to include services setup such as Redis, Postgres, and Nats.

These changes are made to automate the building and deploying processes of the application, ensuring every commit is tested in an environment similar to production. It provides repeatability of the build process and reduces manual errors. Also, the specific tagging of Docker images allows for easier tracking and management of different versions of the application.