chore: 🧑‍💻 add developer dashboard
This commit introduces a UI dashboard (flame), themes and configuration for databases in the Docker environment. It also includes changes in the README formatting and adding `.sqlite`, `db_backups` and `uploads` to `.gitignore`.  The dashboard provides an easier navigation between services related to the repository. This update improves usability and provides easier and customized control of the application's environment, aligned with its specific needs and aesthetics.