feat: ✨ add `Mailer` abstraction
Created new mailer classes for server application. The classes include EmailMessage, EmailAttachment, Mailer, EmailAddress, Html, and PlainText. The Mailer class is an abstract class that will send an EmailMessage. EmailMessage includes key information for sending an email such as recipient and sender details, email's subject and body, etc. It acts as a structure for an email. EmailAddress, Html, and PlainText are the basic types used for forming an EmailMessage.