refactor: ♻️ update project structure and dependencies
Deleted-related files to a top-level account directory and updated their import paths to reflect this change. Also created a new PaymentMethodController class in the Payment module, and updated the version of some nested dependencies in package-lock.yaml. This structural changes were done to keep the application modular and less coupled, the new controller was added as a requirement for new functionalities and dependencies were updated due to a needed update.