refactor: 🔥 remove custom identification library
Removed a custom library that was used for generating unique identifiers for the system's entities. Functions like cuid, uuid or typeid were becoming obsolete due to the introduction of up-to-date identification libraries like ksuid or ulid. The custom library was unoptimized and added unnecessary complexity to the system. Migrating to the new libraries will simplify the codebase and increase performance.