ci: update server workflows
The commit updates the Continuous Integration workflow to look for 'main' branch and renames it with an emoji. The 'cd.yml' file has been updated to monitor the renamed CI workflow. A README file has also been added in 'account domain services', outlining the various features that are supported by the account related services.