refactor: ♻️ remove authorization entities, add checkout entities
This commit deletes 'Permission' and 'Role' classes from the authorization module. It adds 'CheckoutStatus' and 'CheckoutItem' to the checkout module, and renames the 'AccountDto' in the account module to 'AccountViewModel'. An openapi3.json file has also been updated, and a new README file for the sale channel module has been added.