feat: add role and group controllers, create Grant and Role services
This refactor introduces new Role and Group controllers in the authorization module. These controllers handle CRUD operations for roles and groups, respectively. In addition to the controllers, this update also added the GrantService and RoleService abstract classes. Changes are also made in the structure of GitHub actions and 'session-controller.ts'. The structure of Prisma schema is modified and renamed some models.