Update server details and docker image in Terraform scripts
The Terraform scripts have been updated to display application's domain names in development server details. The docker image has also been updated to use the latest version instead of the dev version. This commit is aimed at improving the accuracy and relevance of server information, as well as using up-to-date docker images for the application deployment.