refactor: ♻️ update structure of `CacheManagerModule`
Refactored the code by re-organizing the file structure for cache manager and rewriting parts of the logger module. In the cache manager, the main class was moved to a new 'provider' subdirectory and a 'contract' subdirectory was added for the cache manager interfaces. In the logger module, redundant codes were removed to tidy up the interface. Lastly, unnecessary README files were moved and replaced with updated entity files. These changes aim to make the codebase cleaner and easier to navigate, and thus more maintainable.