refactor: ♻️ add validation for Longitude
"prepare" script was enabled in the server package.json file to ensure necessary tasks run during preparation stage. Additionally, a detailed validation was introduced to the Longitude type in 'geoseeker' library to carefully control and validate the input. Enabled several transformations in the tsconfig file to improve the operation of the server application.