refactor: ♻️ geoseeker library and update coordinate models
Deleted 'point.ts', renamed and updated 'coordinates.ts', and added new fields and types in 'altitude.ts' and 'geohash.ts'. These changes are part of a larger refactor for our geoseeker library where geo-coordinates and related fields are being moved to the value-objects directory. The update involved introducing Altitude and Geohash types in coordinates and removal of the unneeded Coordinates reference from the address model.