refactor: 🔥 token-related files and refactor structure
Deleted files related to JSON Web Tokens in the authentication module and created new similar files in the token module. The aim of this change was to make the token files more autonomous, thus improving code organization and project structure. Also made minor modifications in the session module to correctly locate the shifted token files.