feat: ✨ add MessageType enum and apply on message classes
Implemented a new 'MessageType' enumeration in a new file and applied this enum class on 'message', 'command', and 'event' classes. This enhances code cleanliness and reduces redundancy by avoiding the re-declaration of the same specific string type set in each of these classes. The enum usage provides a central control for the types.