refactor: ♻️ `Longitude` type to use Opaque in geoseeker library
This commit refactors Longitude type in the geoseeker library. Instead of using a custom validateLongitude function, it now utilizes the Opaque type and the UnwrapOpaque utility from the opaque.js file in the common libraries directory.

This simplifies the code and takes advantage of existing type validation functions, thus improving maintainability and code clarity.