feat: ✨ add "No-operation Notification Service" for sending notifications
This update adds a "No-operation" implementation of the previously abstract class, `NotificationService`. The newly added service, `NoopNotificationService`, creates log when implementing the `sendNotification` method inheriting from `NotificationService`. An important impact of this change is the new capability to build the application without the need for a specific notification system implementation. This is especially useful for offline development and testing. Parts of the application rely on the previous abstract service and can now function without throwing errors for missing required methods.