refactor: ♻️ code to maximize Domain-Driven Design principles
In this refactor, `DomainEvent` has been utilized more effectively as a domain-focused event factory. A key change is the integration of an `AccountStatus` enum to bring more clarity and precision into the account status definition. This commit also enhances the OpenAPI documentation by adding a health check endpoint. Efficiency in the email verification process has been improved by a massive tidying-up of the related code. This is an essential step to align closer with Domain-Driven Design principles.