refactor: ♻️ authentication and configuration files, update API docs
This commit primarily includes refactor of authentication-related files, updates to configuration files, as well as changes to the API documentation. It also involves addition of files for transactional outbox and cache replacement policy along with edits to username.ts and the Docker setup files. The commit aims to streamline and improve the existing codes for better functionality and easier understanding and use.